requirements.txt
Package                   Version
------------------------- --------------------
altair                    5.1.2
attrs                     23.1.0
blinker                   1.7.0
cachetools                5.3.2
certifi                   2023.7.22
chardet                   4.0.0
charset-normalizer        3.3.2
click                     8.1.7
colorama                  0.4.6
contourpy                 1.2.0
cycler                    0.10.0
filelock                  3.13.1
fonttools                 4.44.0
fsspec                    2023.10.0
gitdb                     4.0.11
GitPython                 3.1.40
idna                      2.10
importlib-metadata        6.8.0
Jinja2                    3.1.2
jsonschema                4.19.2
jsonschema-specifications 2023.7.1
kiwisolver                1.4.5
markdown-it-py            3.0.0
MarkupSafe                2.1.3
matplotlib                3.8.1
mdurl                     0.1.2
mpmath                    1.3.0
networkx                  3.2.1
numpy                     1.26.1
opencv-python             4.8.1.78
opencv-python-headless    4.8.0.74
packaging                 23.2
pandas                    2.1.2
Pillow                    10.1.0
pip                       23.3
protobuf                  4.25.0
psutil                    5.9.6
py-cpuinfo                9.0.0
pyarrow                   14.0.1
pydeck                    0.8.1b0
Pygments                  2.16.1
pyparsing                 3.1.1
python-dateutil           2.8.2
pytz                      2023.3.post1
PyYAML                    6.0.1
referencing               0.30.2
requests                  2.31.0
requests-toolbelt         1.0.0
rich                      13.6.0
roboflow                  1.1.9
rpds-py                   0.12.0
scipy                     1.11.3
seaborn                   0.13.0
setuptools                68.0.0
six                       1.16.0
smmap                     5.0.1
streamlit                 1.28.1
supervision               0.16.0
sympy                     1.12
tenacity                  8.2.3
thop                      0.1.1.post2209072238
toml                      0.10.2
toolz                     0.12.0
torch                     2.1.0
torchvision               0.16.0
tornado                   6.3.3
tqdm                      4.66.1
typing_extensions         4.8.0
tzdata                    2023.3
tzlocal                   5.2
ultralytics               8.0.208
urllib3                   2.0.7
validators                0.22.0
watchdog                  3.0.0
wheel                     0.41.2
zipp                      3.17.0
